# perf(jsx): skip the special behavior when the element is in the head.