@yusukebe 2024-09-01
Hey @usualoma 

> The following elements are no longer special behavior when they are already in the head element.

That's a good improvement.

My score is below. Looks good!

```
cpu: Apple M1 Pro
runtime: node v22.6.0 (arm64-darwin)

benchmark        time (avg)             (min … max)       p75       p99      p999
--------------------------------------------------- -----------------------------
• Real World Scenario
--------------------------------------------------- -----------------------------
KitaJS/Html     483 µs/iter     (389 µs … 1'167 µs)    420 µs  1'044 µs  1'137 µs
Typed Html    1'791 µs/iter   (1'573 µs … 2'239 µs)  2'046 µs  2'200 µs  2'239 µs
VHtml         2'182 µs/iter   (2'045 µs … 2'770 µs)  2'257 µs  2'733 µs  2'770 µs
React JSX     5'994 µs/iter   (5'551 µs … 6'866 µs)  6'437 µs  6'866 µs  6'866 µs
Preact          970 µs/iter     (656 µs … 4'845 µs)    892 µs  2'825 µs  4'845 µs
React         6'203 µs/iter  (5'647 µs … 11'046 µs)  6'681 µs 11'046 µs 11'046 µs
Common Tags   2'965 µs/iter   (2'762 µs … 3'621 µs)  2'890 µs  3'583 µs  3'621 µs
Ghtml           221 µs/iter     (170 µs … 1'620 µs)    208 µs  1'180 µs  1'499 µs
JSXTE         3'869 µs/iter   (3'745 µs … 4'091 µs)  3'902 µs  4'076 µs  4'091 µs
hono/jsx      1'075 µs/iter     (959 µs … 1'867 µs)  1'003 µs  1'779 µs  1'867 µs

summary for Real World Scenario
  Ghtml
   2.19x faster than KitaJS/Html
   4.39x faster than Preact
   4.87x faster than hono/jsx
   8.11x faster than Typed Html
   9.87x faster than VHtml
   13.42x faster than Common Tags
   17.51x faster than JSXTE
   27.13x faster than React JSX
   28.07x faster than React
```